update common services responses

+++ src/main/java/com/btv/app/features/book/services/BookController.java
@@ -1,70 +1,73 @@
 package com.btv.app.features.book.services;
 
 import com.btv.app.features.book.model.Book;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 @RestController
-@RequestMapping("api/book")
+@RequestMapping("api")
+@AllArgsConstructor
 public class BookController {
     private final BookService bookService;
-
-    @Autowired
-    public BookController(BookService bookService) {
-        this.bookService = bookService;
-    }
-
     @GetMapping("/getAllBooks")
-    public List<Book> getAllBooks(){
+    public ResponseEntity<List<Book>> getAllBooks(){
         try {
-            return bookService.getAllBooks();
+            List<Book> res = bookService.getAllBooks();
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @GetMapping("/getBookByID/{id}")
-    public Book getBookByID(@PathVariable("id") Long id){
+    public ResponseEntity<Book> getBookByID(@PathVariable("id") Long id){
         try{
-            return bookService.getBookByID(id);
+            Book res = bookService.getBookByID(id);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
-    @PostMapping("/addBook")
-    public Book addBook(@ModelAttribute Book book){
+    @PostMapping("admin/addBook")
+    public ResponseEntity<Book> addBook(@ModelAttribute Book book){
         try{
-            Book tmp = bookService.addBook(book);
-            return new Book(tmp);
+            Book res = bookService.addBook(book);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
-    @PutMapping("/modifyBook/{id}")
-    public Book modifyBook(@PathVariable("id") Long id, @ModelAttribute Book book){
+    @PutMapping("admin/modifyBook/{id}")
+    public ResponseEntity<Book> modifyBook(@PathVariable("id") Long id, @ModelAttribute Book book){
         try{
             Book curBook = bookService.getBookByID(id);
+            if(curBook == null){
+                return ResponseEntity.status(404).build();
+            }
             Book tmp = bookService.modifyBook(curBook, book);
-            return new Book(tmp);
+            return ResponseEntity.status(200).body(tmp);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
-    @DeleteMapping("/deleteBook/{id}")
-    public void deleteBook(@PathVariable("id") Long id){
+    @DeleteMapping("admin/deleteBook/{id}")
+    public ResponseEntity<Book> deleteBook(@PathVariable("id") Long id){
         try{
+            Book curBook = bookService.getBookByID(id);
+            if(curBook == null){
+                return ResponseEntity.status(404).build();
+            }
             bookService.deleteBook(id);
+            return ResponseEntity.status(200).body(curBook);
         } catch (Exception e) {
-            System.out.println(e);
+            return ResponseEntity.status(500).build();
         }
     }
 }
