+++ src/main/java/com/btv/app/features/user/services/UserController.java
@@ -1,8 +1,10 @@
 package com.btv.app.features.user.services;
 
 import com.btv.app.features.user.models.User;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,70 +12,63 @@ import java.util.ArrayList;
 import java.util.List;
 @RestController
 @RequestMapping("api")
+@AllArgsConstructor
 public class UserController {
     private final UserService userService;
-
-    @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
     @GetMapping("admin/getAllUsers")
-    public List<User> getAllUsers(){
+    public ResponseEntity<List<User>> getAllUsers(){
         try {
-            List<User> UserList =  userService.getAllUsers();
-            List<User> resList = new ArrayList<User>();
-
-            for(User user: UserList){
-                User tmp = new User(user);
-                resList.add(tmp);
-            }
-
-            return resList;
+            List<User> res = userService.getAllUsers();
+            ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
     @GetMapping("admin/getUserByID/{id}")
-    public User getUserByID(@PathVariable("id") Long id){
+    public ResponseEntity<User> getUserByID(@PathVariable("id") Long id){
         try{
-            User tmp = userService.getUserByID(id);
-            return new User(tmp);
+            User res = userService.getUserByID(id);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @PostMapping("admin/addUser")
-        public User addUser(@ModelAttribute User user){
+        public ResponseEntity<User> addUser(@ModelAttribute User user){
         try{
-            User tmp = userService.addUser(user);
-            return new User(tmp);
+            User res = userService.addUser(user);
+            return ResponseEntity.status(200).body(res);
         } catch (DataIntegrityViolationException e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @PutMapping("admin/modifyUser/{id}")
-    public User modifyUser(@PathVariable("id") Long id, @ModelAttribute User user){
+    public ResponseEntity<User> modifyUser(@PathVariable("id") Long id, @ModelAttribute User user){
         try{
             User curUser = userService.getUserByID(id);
-            User tmp = userService.modifyUser(curUser, user);
-            return new User(tmp);
+            if(curUser == null){
+                return ResponseEntity.status(404).build();
+            }
+            User res = userService.modifyUser(curUser, user);
+            return ResponseEntity.status(200).body(res);
         } catch (Exception e) {
-            System.out.println(e);
-            return null;
+            return ResponseEntity.status(500).build();
         }
     }
 
     @DeleteMapping("admin/deleteUser/{id}")
-    public void deleteUser(@PathVariable("id") Long id){
+    public ResponseEntity<User> deleteUser(@PathVariable("id") Long id){
         try{
+            User curUser = userService.getUserByID(id);
+            if(curUser == null){
+                return ResponseEntity.status(404).build();
+            }
             userService.deleteUser(id);
+            return ResponseEntity.status(200).body(curUser);
         } catch (Exception e) {
-            System.out.println(e);
+            return ResponseEntity.status(500).build();
         }
     }
